Allow building with Visual Studio up to 2022
This commits hardcodes that the toolset should be vc14 also when
opening the solution with IDE versions higher than 14.0 (i.e. 2015).

In the future, to allow building with both vc14 and newer compilers,
we should probably create ad-hoc configurations in QuantLib.vcxproj.
The alternative would be to have one clone of the project file
for each toolset, as is done in QuantlibAddin components.